Moved SafeJson classes to separate files
