+++ Source/Libraries/NetCore/NetCore.csproj
@@ -61,6 +61,10 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="SafeJsonTypeSerialization\JsonKnownTypesBinder.cs" />
+    <Compile Include="SafeJsonTypeSerialization\TypeWrapper.cs" />
+    <Compile Include="SafeJsonTypeSerialization\UntypedToTypedValueContractResolver.cs" />
+    <Compile Include="SafeJsonTypeSerialization\UntypedToTypedValueConverter.cs" />
     <Compile Include="AllSpec.cs" />
     <Compile Include="Connector.cs" />
     <Compile Include="NetCore_Extensions.cs" />
